darwin: fix topology
The parent device should always already be in the context when a
device is processed. Use the parent device's location to find the
parent dev.

Note: we might want to replicate the notion of a locationID in the
core. If done right the parent dev could be determined in the core.